fix: Fix circular imports when importing collectionName from model, also refactor interfaces of model to be in separated file
