+++ src/accessToken/accessToken.interface.ts
@@ -2,13 +2,13 @@ import { IBaseModel } from '../generic/generic.interface';
 import { IClient } from '../client/client.interface';
 import { IUser } from '../user/user.interface';
 
-export interface IAuthCode extends IBaseModel {
-  value: string;
+export interface IAccessToken extends IBaseModel {
   clientId: string | IClient; // Client ID or Client Model after population
   userId: string | IUser; // User ID or User model after population
-  redirectUri: string;
+  value: string;
   scopes: [string];
+  grantType: string;
   expireAt: Date;
 }
 
-export const collectionName = 'AuthCode';
+export const collectionName = 'AccessToken';
