+++ src/accessToken/accessToken.validator.ts
@@ -1,5 +1,5 @@
 import { refValidator, uniqueValidator } from '../generic/generic.validator';
-import { collectionName } from './accessToken.model';
+import { collectionName } from './accessToken.interface';
 
 // Access Token reference validator
 export const accessTokenRefValidator = [
