+++ src/authCode/authCode.validator.ts
@@ -1,6 +1,6 @@
 import { refValidator, uniqueValidator } from '../generic/generic.validator';
 import { collectionName } from './authCode.interface';
-console.log(collectionName);
+
 // Auth Code reference validator
 export const authCodeRefValidator = [
   refValidator.bind({}, collectionName, '_id'),
