+++ src/client/client.validator.ts
@@ -1,5 +1,5 @@
 import { refValidator, uniqueValidator } from '../generic/generic.validator';
-import { collectionName } from './client.model';
+import { collectionName } from './client.interface';
 
 // Client reference validator
 export const clientRefValidator = [
