+++ src/refreshToken/refreshToken.model.ts
@@ -1,18 +1,12 @@
 import { Schema, model } from 'mongoose';
-import { IBaseModel } from '../generic/generic.interface';
+import { IRefreshToken, collectionName } from './refreshToken.interface';
 import { refreshTokenUniqueValueValidator } from './refreshToken.validator';
 import { accessTokenRefValidator } from '../accessToken/accessToken.validator';
 import {
-  collectionName as AccessTokenModelName,
   IAccessToken,
-} from '../accessToken/accessToken.model';
+  collectionName as AccessTokenModelName,
+} from '../accessToken/accessToken.interface';
 import config from '../config';
-
-export interface IRefreshToken extends IBaseModel {
-  value: string;
-  accessTokenId: string | IAccessToken; // Access Token id or Access Token after population
-  expiresAt: Date;
-}
 
 // TODO: Check if multiple refresh token reference the same access token
 
@@ -37,7 +31,6 @@ const refreshTokenSchema = new Schema({
   },
 });
 
-export const collectionName = 'RefreshToken';
 const refreshTokenModel = model<IRefreshToken>(collectionName, refreshTokenSchema);
 
 export default refreshTokenModel;
