+++ src/user/user.model.ts
@@ -1,13 +1,7 @@
 import { Schema, model } from 'mongoose';
-import { IBaseModel } from '../generic/generic.interface';
+import { IUser, collectionName } from './user.interface';
 import { userUniqueEmailValidator } from './user.validator';
 import { generatePasswordHash } from '../utils/hashUtils';
-
-export interface IUser extends IBaseModel {
-  name: string;
-  email: string;
-  password: string;
-}
 
 const userSchema = new Schema({
   name: {
@@ -30,7 +24,6 @@ userSchema.pre<IUser>('save', function save() {
   this.password = generatePasswordHash(this.password);
 });
 
-export const collectionName = 'User';
 const userModel = model<IUser>(collectionName, userSchema);
 
 export default userModel;
